make tests package names consistent with the folder structure   - extract runSwaggerSpec method from a package object into a trait   - remove unnecessary imports
