--- src/test/scala/generators/AkkaHttp/Client/DefaultParameters.scala
@@ -1,175 +1,0 @@
-package tests.generators.akkaHttp.client
-
-import _root_.io.swagger.parser.SwaggerParser
-import cats.instances.all._
-import com.twilio.swagger._
-import com.twilio.guardrail.generators.AkkaHttp
-import com.twilio.guardrail.{Client, Clients, Context, ClientGenerator, CodegenApplication, ProtocolGenerator, Target}
-import org.scalatest.{FunSuite, Matchers}
-
-class DefaultParametersTest extends FunSuite with Matchers {
-  import scala.meta._
-
-  val swagger = s"""
-    |swagger: '2.0'
-    |host: petstore.swagger.io
-    |paths:
-    |  "/store/order/{order_id}":
-    |    get:
-    |      tags:
-    |      - store
-    |      x-scala-package: store
-    |      operationId: getOrderById
-    |      produces:
-    |      - application/xml
-    |      - application/json
-    |      parameters:
-    |      - name: order_id
-    |        in: path
-    |        required: true
-    |        type: integer
-    |        maximum: 10
-    |        minimum: 1
-    |        format: int64
-    |      - name: HeaderMeThis
-    |        in: header
-    |        type: string
-    |        required: true
-    |      - name: defparm_opt
-    |        in: query
-    |        type: integer
-    |        format: int32
-    |        default: 1
-    |      - name: defparm
-    |        in: query
-    |        type: integer
-    |        format: int32
-    |        required: true
-    |        default: 2
-    |      responses:
-    |        '200':
-    |          description: successful operation
-    |          schema:
-    |            "$$ref": "#/definitions/Order"
-    |        '400':
-    |          description: Invalid ID supplied
-    |        '404':
-    |          description: Order not found
-    |    delete:
-    |      tags:
-    |      - store
-    |      x-scala-package: store
-    |      summary: Delete purchase order by ID
-    |      description: For valid response try integer IDs with positive integer value.
-    |        Negative or non-integer values will generate API errors
-    |      operationId: deleteOrder
-    |      produces:
-    |      - application/xml
-    |      - application/json
-    |      parameters:
-    |      - name: order_id
-    |        in: path
-    |        description: ID of the order that needs to be deleted
-    |        required: true
-    |        type: integer
-    |        minimum: 1
-    |        format: int64
-    |      responses:
-    |        '400':
-    |          description: Invalid ID supplied
-    |        '404':
-    |          description: Order not found
-    |securityDefinitions:
-    |  petstore_auth:
-    |    type: oauth2
-    |    authorizationUrl: http://petstore.swagger.io/oauth/dialog
-    |    flow: implicit
-    |    scopes:
-    |      write:pets: modify pets in your account
-    |      read:pets: read your pets
-    |  api_key:
-    |    type: apiKey
-    |    name: api_key
-    |    in: header
-    |definitions:
-    |  Order:
-    |    type: object
-    |    properties:
-    |      id:
-    |        type: integer
-    |        format: int64
-    |      petId:
-    |        type: integer
-    |        format: int64
-    |      quantity:
-    |        type: integer
-    |        format: int32
-    |      shipDate:
-    |        type: string
-    |        format: date-time
-    |      status:
-    |        type: string
-    |        description: Order Status
-    |        enum:
-    |        - placed
-    |        - approved
-    |        - delivered
-    |      complete:
-    |        type: boolean
-    |        default: false
-    |    xml:
-    |      name: Order
-    |""".stripMargin
-
-  test("Ensure responses are generated") {
-    val (
-      _,
-      Clients(Client(tags, className, statements) :: _),
-      _
-    ) = runSwaggerSpec(swagger)(Context.empty, AkkaHttp)
-
-    tags should equal (Seq("store"))
-
-    val List(cmp, cls) = statements.dropWhile(_.isInstanceOf[Import])
-
-    val companion = q"""
-    object StoreClient {
-      def apply(host: String = "http://petstore.swagger.io")(implicit httpClient: HttpRequest => Future[HttpResponse], ec: ExecutionContext, mat: Materializer): StoreClient =
-        new StoreClient(host = host)(httpClient = httpClient, ec = ec, mat = mat)
-      def httpClient(httpClient: HttpRequest => Future[HttpResponse], host: String = "http://petstore.swagger.io")(implicit ec: ExecutionContext, mat: Materializer): StoreClient =
-        new StoreClient(host = host)(httpClient = httpClient, ec = ec, mat = mat)
-    }
-    """
-
-    val client = q"""
-    class StoreClient(host: String = "http://petstore.swagger.io")(implicit httpClient: HttpRequest => Future[HttpResponse], ec: ExecutionContext, mat: Materializer) {
-      val basePath: String = ""
-      private[this] def wrap[T: FromEntityUnmarshaller](resp: Future[HttpResponse]): EitherT[Future, Either[Throwable, HttpResponse], T] = {
-        EitherT(resp.flatMap(resp => if (resp.status.isSuccess) {
-          Unmarshal(resp.entity).to[T].map(Right.apply _)
-        } else {
-          FastFuture.successful(Left(Right(resp)))
-        }).recover({
-          case e: Throwable =>
-            Left(Left(e))
-        }))
-      }
-      def getOrderById(orderId: Long, defparmOpt: Option[Int] = Option(1), defparm: Int = 2, headerMeThis: String, headers: scala.collection.immutable.Seq[HttpHeader] = Nil): EitherT[Future, Either[Throwable, HttpResponse], Order] = {
-        val allHeaders = headers ++ scala.collection.immutable.Seq[Option[HttpHeader]](Some(RawHeader("HeaderMeThis", Formatter.show(headerMeThis)))).flatten
-        wrap[Order](Marshal(HttpEntity.Empty).to[RequestEntity].flatMap { entity =>
-          httpClient(HttpRequest(method = HttpMethods.GET, uri = host + basePath + "/store/order/" + Formatter.addPath(orderId) + "?" + Formatter.addArg("defparm_opt", defparmOpt) + Formatter.addArg("defparm", defparm), entity = entity, headers = allHeaders))
-        })
-      }
-      def deleteOrder(orderId: Long, headers: scala.collection.immutable.Seq[HttpHeader] = Nil): EitherT[Future, Either[Throwable, HttpResponse], IgnoredEntity] = {
-        val allHeaders = headers ++ scala.collection.immutable.Seq[Option[HttpHeader]]().flatten
-        wrap[IgnoredEntity](Marshal(HttpEntity.Empty).to[RequestEntity].flatMap { entity =>
-          httpClient(HttpRequest(method = HttpMethods.DELETE, uri = host + basePath + "/store/order/" + Formatter.addPath(orderId), entity = entity, headers = allHeaders))
-        })
-      }
-    }
-    """
-
-    cmp.structure should equal(companion.structure)
-    cls.structure should equal(client.structure)
-  }
-}
