+++ src/test/scala/support/ScalaMetaMatchers.scala
@@ -8,8 +8,8 @@ trait ScalaMetaMatchers {
     def apply(left: Term): MatchResult = {
       MatchResult(
         left.structure == right.structure,
-        s"""${left} did not match structure ${right}""",
-        s"""${left} matched structure ${right}"""
+        s"""$left did not match structure $right""",
+        s"""$left matched structure $right"""
       )
     }
   }
