+++ src/test/scala/tests/generators/akkaHttp/client/contentType/TextPlainTest.scala
@@ -1,0 +1,75 @@
+package tests.generators.akkaHttp.client.contentType
+
+import com.twilio.guardrail.generators.AkkaHttp
+import com.twilio.guardrail.{Client, Clients, Context}
+import org.scalatest.{FunSuite, Matchers}
+import support.SwaggerSpecRunner
+
+import scala.meta._
+
+class TextPlainTest extends FunSuite with Matchers with SwaggerSpecRunner{
+  val swagger: String = s"""
+    |swagger: "2.0"
+    |info:
+    |  title: Whatever
+    |  version: 1.0.0
+    |host: localhost:1234
+    |schemes:
+    |  - http
+    |paths:
+    |  /foo:
+    |    put:
+    |      operationId: putFoo
+    |      consumes:
+    |        - text/plain
+    |      parameters:
+    |        - name: body
+    |          in: body
+    |          required: true
+    |          schema:
+    |            type: string
+    |      responses:
+    |        200:
+    |          description: Success
+    |""".stripMargin
+
+  test("Properly handle all methods") {
+    val (
+      _,
+      Clients(Client(tags, className, statements) :: _),
+      _
+    ) = runSwaggerSpec(swagger)(Context.empty, AkkaHttp)
+    val List(cmp, cls) = statements.dropWhile(_.isInstanceOf[Import])
+
+    val companion = q"""
+      object Client {
+        def apply(host: String = "http://localhost:1234")(implicit httpClient: HttpRequest => Future[HttpResponse], ec: ExecutionContext, mat: Materializer): Client = new Client(host = host)(httpClient = httpClient, ec = ec, mat = mat)
+        def httpClient(httpClient: HttpRequest => Future[HttpResponse], host: String = "http://localhost:1234")(implicit ec: ExecutionContext, mat: Materializer): Client = new Client(host = host)(httpClient = httpClient, ec = ec, mat = mat)
+      }
+    """
+    val client = q"""
+      class Client(host: String = "http://localhost:1234")(implicit httpClient: HttpRequest => Future[HttpResponse], ec: ExecutionContext, mat: Materializer) {
+        val basePath: String = ""
+        private[this] def wrap[T: FromEntityUnmarshaller](resp: Future[HttpResponse]): EitherT[Future, Either[Throwable, HttpResponse], T] = {
+          EitherT(resp.flatMap(resp => if (resp.status.isSuccess) {
+            Unmarshal(resp.entity).to[T].map(Right.apply _)
+          } else {
+            FastFuture.successful(Left(Right(resp)))
+          }).recover({
+            case e: Throwable =>
+              Left(Left(e))
+          }))
+        }
+        def putFoo(body: String, headers: scala.collection.immutable.Seq[HttpHeader] = Nil): EitherT[Future, Either[Throwable, HttpResponse], IgnoredEntity] = {
+          val allHeaders = headers ++ scala.collection.immutable.Seq[Option[HttpHeader]]().flatten
+          wrap[IgnoredEntity](Marshal(TextPlain(body)).to[RequestEntity].flatMap {
+            entity => httpClient(HttpRequest(method = HttpMethods.PUT, uri = host + basePath + "/foo", entity = entity, headers = allHeaders))
+          })
+        }
+      }
+    """
+
+    cmp.structure should equal(companion.structure)
+    cls.structure should equal(client.structure)
+  }
+}
